+++ contracts/zero-ex/CHANGELOG.json
@@ -1,5 +1,14 @@
 [
     {
+        "version": "0.2.0",
+        "changes": [
+            {
+                "note": "Export migration tools",
+                "pr": 2612
+            }
+        ]
+    },
+    {
         "timestamp": 1592969527,
         "version": "0.1.1",
         "changes": [
