+++ contracts/zero-ex/src/wrappers.ts
@@ -13,6 +13,10 @@ export * from '../generated-wrappers/i_simple_function_registry';
 export * from '../generated-wrappers/i_token_spender';
 export * from '../generated-wrappers/i_transform_erc20';
 export * from '../generated-wrappers/initial_migration';
+export * from '../generated-wrappers/ownable';
 export * from '../generated-wrappers/pay_taker_transformer';
+export * from '../generated-wrappers/simple_function_registry';
+export * from '../generated-wrappers/token_spender';
+export * from '../generated-wrappers/transform_erc20';
 export * from '../generated-wrappers/weth_transformer';
 export * from '../generated-wrappers/zero_ex';
