+++ contracts/zero-ex/tsconfig.json
@@ -13,7 +13,11 @@
         "generated-artifacts/ITokenSpender.json",
         "generated-artifacts/ITransformERC20.json",
         "generated-artifacts/InitialMigration.json",
+        "generated-artifacts/Ownable.json",
         "generated-artifacts/PayTakerTransformer.json",
+        "generated-artifacts/SimpleFunctionRegistry.json",
+        "generated-artifacts/TokenSpender.json",
+        "generated-artifacts/TransformERC20.json",
         "generated-artifacts/WethTransformer.json",
         "generated-artifacts/ZeroEx.json",
         "test/generated-artifacts/AffiliateFeeTransformer.json",
